refactor(webapp): move agent/chat types out of ui/ to owning layers
Resolves the layer inversion in #1280: ui/types.ts held cross-cutting
agent-domain types (AgentHandle, AgentEvent, ChatMessage, ...) that every
lower layer imported UPWARD into ui/.

- Move AgentHandle + AgentEvent to core/agent-types.ts
- Move ChatMessage, ToolCall, Session, MessageRole, LickState to
  scoops/chat-types.ts
- ui/types.ts becomes a re-export shim so the UI keeps a stable path
- Repoint the 6 production back-edges (kernel facade, 4 scoops modules,
  chrome-extension offscreen-bridge) and 3 scoops tests at the new layers

Closes #1280
